+++ src/protocols/utils/dist.h
@@ -35,7 +35,7 @@ struct nn_dist_data {
 };
 
 struct nn_dist {
-    size_t count;
+    uint32_t count;
     struct nn_list pipes;
 };
